requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
beautifulsoup4==4.13.4
readability-lxml==0.8.1
lxml==5.4.0
lxml_html_clean==0.4.2
python-dateutil==2.9.0.post0
pydantic==2.11.7
pydantic-settings==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
